requirements.txt
Django==2.2.10
djangorestframework==3.12.1
django-cors-headers==3.5.0
django-heroku==0.3.1
djongo==1.3.6
gunicorn==20.0.4 
whitenoise==5.2.0
opencv-python
urllib3==1.25.11
numpy==1.19.5
psycopg2==2.8.6
pymongo==3.12.3
pytz==2020.1
sqlparse==0.2.4
dj-database-url==0.5.0 

